+++ packages/ui-website/components/Layout/index.tsx
@@ -5,49 +5,9 @@ import { Router, useRouter } from 'next/router';
 import Styled from './styles';
 
 import Header from './Header';
-import Aside from './Aside';
+import Main from './Main';
 import PlaceholderPage from './PlaceholderPage';
 import Footer from './Footer';
-
-const Main: React.FC<{ pathname: string }> = ({ pathname, children }) => {
-  const [isOpen, setOpen] = useState(false);
-
-  useEffect(() => {
-    Router.events.on('routeChangeStart', () => {
-      setOpen(false);
-    });
-
-    return () => {
-      Router.events.off('routeChangeStart', () => {
-        setOpen(false);
-      });
-    };
-  }, []);
-
-  if (/^\/components\/*/.test(pathname)) {
-    return (
-      <>
-        <Aside />
-
-        <Styled.ComponentsButton isOpen={isOpen} onClick={() => setOpen(!isOpen)}>
-          <Styled.ComponentsIcon />
-        </Styled.ComponentsButton>
-
-        <Styled.ComponentsDrawer showOverlay isOpen={isOpen} onChange={() => setOpen(false)} position="right">
-          <Styled.MobileAside />
-        </Styled.ComponentsDrawer>
-
-        <Styled.ComponentsMain>{children}</Styled.ComponentsMain>
-      </>
-    );
-  }
-
-  if (/^\/docs\/*/.test(pathname)) {
-    return <Styled.DocsMain>{children}</Styled.DocsMain>;
-  }
-
-  return <Styled.IndexMain>{children}</Styled.IndexMain>;
-};
 
 const Layout: FC = ({ children }) => {
   const [isLoading, setLoading] = useState(false);
